View and controller for LPI
