+++ app/controllers/CalculationController.scala
@@ -80,14 +80,9 @@ class CalculationController @Inject()(authenticate: AuthenticationPredicate,
     financialDetailsService.getFinancialDetails(taxYear, user.nino) flatMap {
       case financialDetails@FinancialDetailsModel(documentDetails, _) =>
         val documentDetailsWithDueDates: List[DocumentDetailWithDueDate] = {
-          documentDetails.filter(_.paymentLot.isEmpty).map(
-            documentDetail => DocumentDetailWithDueDate(documentDetail, financialDetails.getDueDateFor(documentDetail)))
+          documentDetails.filter(_.paymentLot.isEmpty).map(documentDetail => DocumentDetailWithDueDate(documentDetail, financialDetails.getDueDateFor(documentDetail)))
         }
-        val documentDetailsWithDueDatesForLpi: List[DocumentDetailWithDueDate] = {
-          documentDetails.filter(_.paymentLot.isEmpty).filter(_.latePaymentInterestAmount.isDefined).map(
-            documentDetail => DocumentDetailWithDueDate(documentDetail, documentDetail.interestEndDate, true))
-        }
-        f((documentDetailsWithDueDates ++ documentDetailsWithDueDatesForLpi))
+        f(documentDetailsWithDueDates)
       case FinancialDetailsErrorModel(NOT_FOUND, _) => f(List.empty)
       case _ => Future.successful(itvcErrorHandler.showInternalServerError())
     }
