+++ app/views/agent/ChargeSummary.scala.html
@@ -22,7 +22,7 @@
 @import models.financialDetails.DocumentDetailWithDueDate
 
 @this(mainTemplate: MainTemplate, backLink: BackLink)
-@(documentDetailWithDueDate: DocumentDetailWithDueDate, chargeHistoryOpt: Option[List[ChargeHistoryModel]], backUrl: String)(implicit request: Request[_], messages: Messages, appConfig: config.FrontendAppConfig, implicitDateFormatter: ImplicitDateFormatter)
+@(documentDetailWithDueDate: DocumentDetailWithDueDate,latePaymentInterestCharge: Boolean, chargeHistoryOpt: Option[List[ChargeHistoryModel]], backUrl: String)(implicit request: Request[_], messages: Messages, appConfig: config.FrontendAppConfig, implicitDateFormatter: ImplicitDateFormatter)
 
 @import java.time.LocalDate
 @import implicitDateFormatter.{longDate, toTaxYearEndDate, toTaxYearStartDate}
@@ -36,32 +36,50 @@
 	</tr>
 }
 
-@mainTemplate(title = messages(s"chargeSummary.${documentDetail.getChargeTypeKey}"),bodyClasses = None){
+@mainTemplate(title = messages(if (latePaymentInterestCharge) s"chargeSummary.lpi.${documentDetail.getChargeTypeKey}" else s"chargeSummary.${documentDetail.getChargeTypeKey}"), bodyClasses = None){
 	@backLink(backUrl)
 
 	<h1 class="heading-xlarge">
 		<span class="heading-secondary">@messages("paymentDue.tax-year", toTaxYearStartDate((documentDetail.taxYear.toInt - 1).toString).toLongDate, toTaxYearEndDate(documentDetail.taxYear).toLongDate)</span>
-		@messages(s"chargeSummary.${documentDetail.getChargeTypeKey}")
+		@messages(if (latePaymentInterestCharge) s"chargeSummary.lpi.${documentDetail.getChargeTypeKey}" else s"chargeSummary.${documentDetail.getChargeTypeKey}")
 	</h1>
 
 	<dl class="govuk-summary-list">
 		<div class="govuk-summary-list__row">
 			<dt class="govuk-summary-list__key">@messages("chargeSummary.dueDate")</dt>
 			<dd class="govuk-summary-list__value">
-				@if(documentDetailWithDueDate.isOverdue && !documentDetail.isPaid) { <span class="govuk-tag govuk-tag--red">@messages("agent.paymentDue.over-due")</span>}
-				@{documentDetailWithDueDate.dueDate.get.toLongDate}
+				@if(documentDetailWithDueDate.isOverdue && !documentDetail.isPaid) {
+				<span class="govuk-tag govuk-tag--red">@messages("agent.paymentDue.over-due")</span>
+				}
+				@{
+				if(latePaymentInterestCharge) documentDetail.interestEndDate.get.toLongDate
+				else documentDetailWithDueDate.dueDate.get.toLongDate
+				}
+			</dd>
+		</div>
+		@if(latePaymentInterestCharge) {
+		<div class="govuk-summary-list__row">
+			<dt class="govuk-summary-list__key">@messages("chargeSummary.lpi.interestPeriod")</dt>
+			<dd class="govuk-summary-list__value">@messages("chargeSummary.lpi.interestPeriod.dates", documentDetail.interestFromDate.get.toLongDateShort, documentDetail.interestEndDate.get.toLongDateShort)</dd>
+		</div>
+		}
+		<div class="govuk-summary-list__row">
+			<dt class="govuk-summary-list__key">@messages("chargeSummary.paymentAmount")</dt>
+			<dd class="govuk-summary-list__value">@{
+				if (latePaymentInterestCharge) documentDetail.latePaymentInterestAmount.get.toCurrency
+				else documentDetail.originalAmount.get.toCurrency
+				}
 			</dd>
 		</div>
 		<div class="govuk-summary-list__row">
-			<dt class="govuk-summary-list__key">@messages("chargeSummary.paymentAmount")</dt>
-			<dd class="govuk-summary-list__value">@documentDetail.originalAmount.get.toCurrency</dd>
-		</div>
-		<div class="govuk-summary-list__row">
 			<dt class="govuk-summary-list__key">@messages("chargeSummary.remainingDue")</dt>
-			<dd class="govuk-summary-list__value">@documentDetail.remainingToPay.toCurrency</dd>
+			<dd class="govuk-summary-list__value">@{
+				if (latePaymentInterestCharge) documentDetail.interestRemainingToPay.toCurrency
+				else documentDetail.remainingToPay.toCurrency
+			}
+			</dd>
 		</div>
 	</dl>
-
 	<p>
 		@messages("chargeSummary.whatYouOwe.textOne")
 		<a id= "what-you-owe-link" href=@controllers.agent.nextPaymentDue.routes.PaymentDueController.show().url>@messages("chargeSummary.whatYouOwe.linkText")</a>
@@ -92,7 +110,19 @@
 					reasonCode = "created",
 					amount = documentDetail.originalAmount
 				)
-
+				@if(latePaymentInterestCharge) {
+				<tr>
+					<td style="white-space: nowrap;">@{
+						documentDetail.interestEndDate.get.toLongDateShort
+						}</td>
+					<td>@messages(s"chargeSummary.lpi.chargeHistory.created.${documentDetail.getChargeTypeKey}")
+					</td>
+					<td class="numeric">@{
+						documentDetail.latePaymentInterestAmount.get.toCurrency
+						}
+					</td>
+				</tr>
+				}
  				@for(charge <- chargeHistory) {
 					@chargeHistoryRow(
 						dateStr = charge.reversalDate,
