--- app/views/templates/GovUkWrapper.scala.html
@@ -1,119 +1,0 @@
-@*
- * Copyright 2021 HM Revenue & Customs
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- *@
-
-@import views.html.layouts.GovUkTemplate
-@import views.html.helpers.phaseBanner
-@import views.html.helpers.recruitmentBannerHelper
-@import play.twirl.api.HtmlFormat
-@import uk.gov.hmrc.play.views.html.layouts.FooterLinks
-@import uk.gov.hmrc.play.views.html.layouts.MainContent
-@import uk.gov.hmrc.play.views.html.helpers.ReportAProblemLink
-@import uk.gov.hmrc.play.views.html.layouts.ServiceInfo
-@import uk.gov.hmrc.play.views.html.layouts.Footer
-@import uk.gov.hmrc.play.views.html.layouts.HeaderNav
-@import uk.gov.hmrc.play.views.html.layouts.HeadWithTrackingConsent
-
-@this(
-        govUkTemplate: GovUkTemplate,
-        mainContentLayout: MainContent,
-        footerLinks: FooterLinks,
-        reportAProblemLink: ReportAProblemLink,
-        serviceInfoLayout: ServiceInfo,
-        footer: Footer,
-        headerNav: HeaderNav,
-        headWithTrackingConsent: HeadWithTrackingConsent
-)
-
-@(title: String,
-  mainClass: Option[String] = None,
-  mainDataAttributes: Option[Html] = None,
-  bodyClasses: Option[String] = None,
-  contentHeader: Option[Html] = None,
-  mainContent: Html = HtmlFormat.empty,
-  serviceInfoContent: Html = HtmlFormat.empty,
-  scriptElem: Option[Html] = None,
-  showLogout: Boolean = true)(implicit request: Request[_], messages: Messages, appConfig: config.FrontendAppConfig)
-
-@linksElem = {
-    <link rel="stylesheet" type="text/css" href='@routes.Assets.at("stylesheets/itvc.css")'>
-    <link rel="stylesheet" type="text/css" href='@routes.Assets.at("stylesheets/contrast.css")'>
-}
-
-@head = {
-    @headWithTrackingConsent(
-        linkElem = Some(linksElem),
-      headScripts = None)
-    <meta name="format-detection" content="telephone=no" />
-}
-
-@headerNavLinks = {
-  @if(showLogout) {<li><a id="sign-out" href="@routes.SignOutController.signOut"
-         data-journey-click="primary-navigation:Click:Signout">@messages("base.sign-out")</a></li>}
-}
-
-@insideHeader = {
-    @headerNav(
-      navTitle = Some(Messages("base.service_name")),
-      navTitleLink = None,
-      showBetaLink = false,
-      navLinks = Some(headerNavLinks))
-}
-
-@afterHeader = {@recruitmentBannerHelper()}
-
-@bodyEnd = {
-    @footer(
-      analyticsToken = None,
-      analyticsHost = "",
-      ssoUrl = None,
-      scriptElem = scriptElem,
-      gaCalls = None)
-}
-
-@footerTop = {}
-@footerLinks = {}
-
-@serviceInfo = {
-    @serviceInfoLayout(
-      betaBanner = views.html.helpers.betaBanner(),
-      includeGridWrapper = false,
-      serviceInfoContent = Some(serviceInfoContent),
-      includeHMRCBranding = true,
-      setLang = messages.lang.code
-    )
-}
-
-@mainContentHeader = {
-
-    @if(contentHeader.isDefined) {
-        <header class="page-header inline-block">@contentHeader.get</header>
-    }
-        @helpers.language_selection()
-}
-
-@getHelpForm = @{reportAProblemLink(appConfig.reportAProblemPartialUrl, appConfig.reportAProblemNonJSUrl)}
-
-@content = {
-    @mainContentLayout(
-      article = mainContent,
-      mainClass = mainClass,
-      mainDataAttributes = mainDataAttributes,
-      mainContentHeader = mainContentHeader,
-      serviceInfo = serviceInfo,
-      getHelpForm = getHelpForm)
-}
-
-@govUkTemplate(Some(messages("titlePattern.serviceName.govUk", title)), bodyClasses)(head, bodyEnd, insideHeader, afterHeader, footerTop, Some(footerLinks), nav=true)(content)
