--- app/views/templates/MainTemplate.scala.html
@@ -1,85 +1,0 @@
-@*
- * Copyright 2021 HM Revenue & Customs
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- *@
-
-@import config.FrontendAppConfig
-@import play.twirl.api.HtmlFormat
-@import uk.gov.hmrc.play.views.html.layouts.Article
-@import views.html.templates.GovUkWrapper
-
-@this(
-        article: Article,
-        govUkWrapper: GovUkWrapper
-)
-
-
-@(title: String,
-  contentHeader: Option[Html] = None,
-  bodyClasses: Option[String] = None,
-  mainClass: Option[String] = None,
-  scriptElem: Option[Html] = None,
-  showLogout: Boolean = true,
-  showBtaHeader: Boolean = true,
-  serviceInfo: Html = Html(""),
-  timeout: Boolean = true
-)(mainContent: Html)(implicit request : Request[_], messages: Messages, appConfig: FrontendAppConfig)
-
-
-@defaultScriptElem = {
-    <link rel="stylesheet" href='@routes.Assets.at("stylesheets/timeout.css")' />
-    <script src='@routes.Assets.at("javascripts/timeout.js")' type="text/javascript"></script>
-    <script>
-        @if(timeout) {
-            $.timeoutDialog({
-                timeout: 900,
-                countdown: 120,
-                keep_alive_url: "@controllers.timeout.routes.SessionTimeoutController.keepAlive().url",
-                messageSeconds: "@messages("timeoutDialog.seconds")",
-                messageMinutes: "@messages("timeoutDialog.minutes")",
-                messageMinute: "@messages("timeoutDialog.minute")",
-                messageMinutesTwo: "@messages("timeoutDialog.minutes.two")",
-                logout_url: "@controllers.timeout.routes.SessionTimeoutController.timeout().url",
-                message: "@messages("timeoutDialog.message")",
-                keep_alive_button_text: "@messages("button.continue")"
-            });
-        }
-        var dialogOpen;
-    </script>
-
-    <script type="text/javascript" src=@routes.Assets.at("javascripts/itvc.js")></script>
-    <script>
-		setAccordianMessages("@messages("sections.open.all")", "@messages("sections.close.all")");
-	</script>
-    <script type="text/javascript" src=@routes.Assets.at("javascripts/skipToMainContent.js")></script>
-    <script type="text/javascript" src=@routes.Assets.at("javascripts/accordian.js")></script>
-    <script type="text/javascript" src=@routes.Assets.at("javascripts/govuk-frontend.js")></script>
-    @scriptElem.fold(HtmlFormat.empty)(x => x)
-}
-
-@mainContentWithGA = {
-    @mainContent
-}
-
-@govUkWrapper(title = title,
-               mainClass = mainClass,
-               bodyClasses = bodyClasses,
-               contentHeader = contentHeader,
-               mainContent = article(mainContentWithGA),
-               serviceInfoContent = serviceInfo,
-               scriptElem = Some(defaultScriptElem),
-               showLogout = showLogout
-)
-
-
