+++ conf/messages
@@ -163,9 +163,6 @@ tax-year-overview.payments.amount                                   = Amount
 tax-year-overview.payments.paymentOnAccount1.text                   = Payment on account 1 of 2
 tax-year-overview.payments.paymentOnAccount2.text                   = Payment on account 2 of 2
 tax-year-overview.payments.balancingCharge.text                     = Remaining balance
-tax-year-overview.payments.lpi.paymentOnAccount1.text               = Late payment interest on payment on account 1 of 2
-tax-year-overview.payments.lpi.paymentOnAccount2.text               = Late payment interest on payment on account 2 of 2
-tax-year-overview.payments.lpi.balancingCharge.text                 = Late payment interest on remaining balance
 tax-year-overview.payments.paid                                     = Paid
 tax-year-overview.payments.unpaid                                   = Unpaid
 tax-year-overview.payments.part-paid                                = Part Paid
