+++ test/controllers/agent/ChargeSummaryControllerSpec.scala
@@ -24,7 +24,7 @@ import config.featureswitch.{ChargeHistory, FeatureSwitching}
 import implicits.ImplicitDateFormatterImpl
 import mocks.auth.MockFrontendAuthorisedFunctions
 import mocks.services.{MockFinancialDetailsService, MockIncomeSourceDetailsService}
-import mocks.views.agent.MockChargeSummary
+import mocks.views.MockChargeSummary
 import models.chargeHistory.ChargeHistoryModel
 import models.financialDetails.FinancialDetailsModel
 import org.mockito.ArgumentMatchers.{any, eq => ameq}
@@ -70,7 +70,9 @@ class ChargeSummaryControllerSpec extends TestSupport
     val currentYear: Int = LocalDate.now().getYear
     val errorHeading = "Sorry, there is a problem with the service"
     val currentFinancialDetails: FinancialDetailsModel = financialDetailsModel(currentYear)
-  }
+		val lateInterestSuccessHeading = "Tax year 6 April 2017 to 5 April 2018 Late payment interest on payment on account 1 of 2"
+
+	}
 
   ".showChargeSummary" when {
 		"getFinancialDetails returns a FinancialDetailsModel" when {
@@ -163,7 +165,7 @@ class ChargeSummaryControllerSpec extends TestSupport
 					.apply(fakeRequestConfirmedClient("AB123456C"))
 
 				status(result) shouldBe OK
-				verify(chargeSummary).apply(any(), chargeHistoryOpt = ameq(Some(chargeHistoryListInAscendingOrder)), any())(any(), any(), any(), any())
+				verify(chargeSummary).apply(any(), any(), chargeHistoryOpt = ameq(Some(chargeHistoryListInAscendingOrder)), any())(any(), any(), any(), any())
 				verify(mockFinancialDetailsService).getChargeHistoryDetails(ameq("XAIT00000000015"), ameq(id1040000123))(any())
 			}
 
@@ -180,7 +182,7 @@ class ChargeSummaryControllerSpec extends TestSupport
 						.apply(fakeRequestConfirmedClient("AB123456C"))
 
 					status(result) shouldBe OK
-					verify(chargeSummary).apply(any(), chargeHistoryOpt = ameq(Some(Nil)), any())(any(), any(), any(), any())
+					verify(chargeSummary).apply(any(), any(), chargeHistoryOpt = ameq(Some(Nil)), any())(any(), any(), any(), any())
 				}
 			}
 
@@ -213,7 +215,7 @@ class ChargeSummaryControllerSpec extends TestSupport
 					.apply(fakeRequestConfirmedClient("AB123456C"))
 
 				status(result) shouldBe OK
-				verify(chargeSummary).apply(any(), chargeHistoryOpt = ameq(None), any())(any(), any(), any(), any())
+				verify(chargeSummary).apply(any(), any(), chargeHistoryOpt = ameq(None), any())(any(), any(), any(), any())
 				verify(mockFinancialDetailsService, never).getChargeHistoryDetails(any(), any())(any())
 			}
 		}
