+++ test/controllers/agent/ConfirmClientUTRControllerSpec.scala
@@ -21,7 +21,7 @@ import config.featureswitch.FeatureSwitching
 import controllers.agent.utils.SessionKeys
 import mocks.MockItvcErrorHandler
 import mocks.auth.MockFrontendAuthorisedFunctions
-import mocks.views.agent.MockConfirmClient
+import mocks.views.MockConfirmClient
 import play.api.http.Status
 import play.api.mvc.MessagesControllerComponents
 import play.api.test.Helpers._
