+++ test/controllers/agent/TaxYearsControllerSpec.scala
@@ -24,7 +24,7 @@ import controllers.Assets.{INTERNAL_SERVER_ERROR, NOT_FOUND, OK, SEE_OTHER}
 import mocks.MockItvcErrorHandler
 import mocks.auth.MockFrontendAuthorisedFunctions
 import mocks.services.{MockCalculationService, MockIncomeSourceDetailsService}
-import mocks.views.agent.MockTaxYears
+import mocks.views.MockTaxYears
 import models.calculation.{Calculation, CalculationErrorModel, CalculationResponseModelWithYear}
 import play.api.mvc.{MessagesControllerComponents, Result}
 import play.api.test.Helpers.{HTML, contentType, defaultAwaitTimeout, redirectLocation}
