+++ test/mocks/views/MockChargeSummary.scala
@@ -14,27 +14,27 @@
  * limitations under the License.
  */
 
-package mocks.views.agent
+package mocks.views
 
 import org.mockito.ArgumentMatchers.any
 import org.mockito.Mockito.{reset, when}
 import org.scalatest.{BeforeAndAfterEach, Suite}
 import org.scalatestplus.mockito.MockitoSugar
 import play.twirl.api.Html
-import views.html.agent.ClientRelationshipFailure
+import views.html.agent.ChargeSummary
 
-trait MockClientRelationshipFailure extends BeforeAndAfterEach with MockitoSugar {
+trait MockChargeSummary extends BeforeAndAfterEach with MockitoSugar {
   self: Suite =>
 
-  val clientRelationshipFailure: ClientRelationshipFailure = mock[ClientRelationshipFailure]
+  val chargeSummary: ChargeSummary = mock[ChargeSummary]
 
   override def beforeEach(): Unit = {
     super.beforeEach()
-    reset(clientRelationshipFailure)
+    reset(chargeSummary)
   }
 
-  def mockClientRelationshipFailure(response: Html): Unit = {
-    when(clientRelationshipFailure.apply(any())(any(), any(), any()))
+  def mockChargeSummary()(response: Html): Unit = {
+    when(chargeSummary.apply(any(), any(), any(),any())(any(), any(), any(), any()))
       .thenReturn(response)
   }
