+++ test/mocks/views/MockTaxCalcBreakdown.scala
@@ -14,27 +14,33 @@
  * limitations under the License.
  */
 
-package mocks.views.agent
+package mocks.views
 
-import org.mockito.ArgumentMatchers.any
+import models.calculation.CalcDisplayModel
+import org.mockito.ArgumentMatchers.{any, eq => matches}
 import org.mockito.Mockito.{reset, when}
 import org.scalatest.{BeforeAndAfterEach, Suite}
 import org.scalatestplus.mockito.MockitoSugar
 import play.twirl.api.Html
-import views.html.agent.ChargeSummary
+import views.html.agent.TaxCalcBreakdown
 
-trait MockChargeSummary extends BeforeAndAfterEach with MockitoSugar {
+trait MockTaxCalcBreakdown extends BeforeAndAfterEach with MockitoSugar {
   self: Suite =>
 
-  val chargeSummary: ChargeSummary = mock[ChargeSummary]
+  val taxCalcBreakdown: TaxCalcBreakdown = mock[TaxCalcBreakdown]
 
   override def beforeEach(): Unit = {
     super.beforeEach()
-    reset(chargeSummary)
+    reset(taxCalcBreakdown)
   }
 
-  def mockChargeSummary()(response: Html): Unit = {
-    when(chargeSummary.apply(any(), any(), any())(any(), any(), any(), any()))
+  def mockTaxCalcBreakdown(taxYear: Int, calcModel: CalcDisplayModel, backUrl: String)
+                       (response: Html): Unit = {
+    when(taxCalcBreakdown.apply(
+      matches(calcModel),
+      matches(taxYear),
+      matches(backUrl)
+    )(any(), any(), any(), any()))
       .thenReturn(response)
   }
